backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.1.1
requests==2.32.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2